backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
ultralytics==8.3.36
torch>=2.0.0
Pillow==11.0.0

